Move Pragma back to package driver.config
