+++ library/driver/src/commonMain/kotlin/io/toxicity/sqlite/mc/driver/config/encryption/EncryptionConfig.kt
@@ -20,8 +20,8 @@ import io.toxicity.sqlite.mc.driver.MCConfigDsl
 import io.toxicity.sqlite.mc.driver.SQLiteMCDriver
 import io.toxicity.sqlite.mc.driver.internal.ext.appendIndent
 import io.toxicity.sqlite.mc.driver.config.FilesystemConfig
-import io.toxicity.sqlite.mc.driver.config.pragma.Pragma
-import io.toxicity.sqlite.mc.driver.config.pragma.MutablePragmas
+import io.toxicity.sqlite.mc.driver.config.Pragma
+import io.toxicity.sqlite.mc.driver.config.MutablePragmas
 import io.toxicity.sqlite.mc.driver.internal.ext.appendColon
 import kotlin.jvm.*
