+++ library/driver/src/commonTest/kotlin/io/toxicity/sqlite/mc/driver/config/encryption/EncryptionConfigUnitTest.kt
@@ -15,8 +15,8 @@
  **/
 package io.toxicity.sqlite.mc.driver.config.encryption
 
-import io.toxicity.sqlite.mc.driver.config.pragma.Pragma
-import io.toxicity.sqlite.mc.driver.config.pragma.mutablePragmas
+import io.toxicity.sqlite.mc.driver.config.Pragma
+import io.toxicity.sqlite.mc.driver.config.mutablePragmas
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
