+++ library/driver/src/jvmAndroidMain/kotlin/io/toxicity/sqlite/mc/driver/internal/JDBCMC4Connection.kt
@@ -20,7 +20,7 @@ import io.matthewnelson.encoding.base64.Base64
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import io.matthewnelson.encoding.core.EncodingException
 import io.matthewnelson.encoding.core.use
-import io.toxicity.sqlite.mc.driver.config.pragma.Pragma
+import io.toxicity.sqlite.mc.driver.config.Pragma
 import org.sqlite.JDBC
 import org.sqlite.jdbc4.JDBC4Connection
 import org.sqlite.jdbc4.JDBC4Statement
