+++ library/gradle-plugin/build.gradle.kts
@@ -15,7 +15,6 @@
  **/
 import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-import org.jetbrains.kotlin.konan.target.HostManager
 
 plugins {
     kotlin("jvm")
